+++ cmd/root.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"net/http"
 	"os"
 
 	"github.com/dynatrace-oss/dtwiz/pkg/logger"
@@ -32,12 +31,13 @@ Set your Dynatrace credentials via environment variables:
 
 Then use dtwiz commands to analyze and instrument your system.`,
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-		debug := debugFlag
-		logger.Init(debug, verbosityFlag)
-		if logger.Verbosity() > 0 {
-			http.DefaultTransport = logger.NewLoggingTransport(http.DefaultTransport)
+		logger.Init(debugFlag, verbosityFlag)
+		switch logger.Verbosity() {
+		case 1:
+			logger.Debug("logging: verbose")
+		case 2:
+			logger.Debug("logging: debug")
 		}
-		logger.Debug("Debug mode enabled", "verbosity", logger.Verbosity())
 	},
 }
 
@@ -72,8 +72,8 @@ func init() {
 		cmd.Help()
 	}
 
-	rootCmd.PersistentFlags().BoolVar(&debugFlag, "debug", false, "enable full HTTP request/response debug logging (equivalent to -vv)")
-	rootCmd.PersistentFlags().CountVarP(&verbosityFlag, "verbose", "v", "verbose output (-v for request/response summary, -vv for full headers and bodies)")
+	rootCmd.PersistentFlags().BoolVar(&debugFlag, "debug", false, "enable debug logging")
+	rootCmd.PersistentFlags().CountVarP(&verbosityFlag, "verbose", "v", "verbose output")
 	rootCmd.PersistentFlags().StringVar(&environmentFlag, "environment", "", "Dynatrace environment URL (also read from DT_ENVIRONMENT)")
 	rootCmd.PersistentFlags().StringVar(&accessTokenFlag, "access-token", "", "Dynatrace API access token (also read from DT_ACCESS_TOKEN)")
 	rootCmd.PersistentFlags().StringVar(&platformTokenFlag, "platform-token", "", "Dynatrace platform token (dt0s16.*) for AWS installer (also read from DT_PLATFORM_TOKEN)")